clean-up of TrackToTrackComparisonHists
