--- DQM/TrackingMonitorSource/plugins/SealModule.cc
@@ -1,8 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "DQM/TrackingMonitorSource/interface/TrackToTrackComparisonHists.h"
-DEFINE_FWK_MODULE(TrackToTrackComparisonHists);
-
-//#include "DQM/TrackingMonitorSource/interface/Vertex2VertexValidator.h"
-//DEFINE_FWK_MODULE(Vertex2VertexValidator);
